cas897: Allow the ClientAuthenticationHandler to become extensible
