+++ cas-server-support-pac4j/src/main/java/org/jasig/cas/support/pac4j/authentication/handler/support/ClientAuthenticationHandler.java
@@ -18,50 +18,24 @@
  */
 package org.jasig.cas.support.pac4j.authentication.handler.support;
 
-
 import org.apache.commons.lang3.StringUtils;
 import org.jasig.cas.authentication.BasicCredentialMetaData;
-import org.jasig.cas.authentication.Credential;
 import org.jasig.cas.authentication.HandlerResult;
 import org.jasig.cas.authentication.PreventedException;
-import org.jasig.cas.authentication.handler.support.AbstractPreAndPostProcessingAuthenticationHandler;
-import org.jasig.cas.authentication.principal.DefaultPrincipalFactory;
-import org.jasig.cas.authentication.principal.PrincipalFactory;
 import org.jasig.cas.support.pac4j.authentication.principal.ClientCredential;
-import org.pac4j.core.client.Client;
 import org.pac4j.core.client.Clients;
-import org.pac4j.core.context.J2EContext;
-import org.pac4j.core.context.WebContext;
-import org.pac4j.core.credentials.Credentials;
 import org.pac4j.core.profile.UserProfile;
-import org.springframework.webflow.context.ExternalContextHolder;
-import org.springframework.webflow.context.servlet.ServletExternalContext;
 
 import javax.security.auth.login.FailedLoginException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.validation.constraints.NotNull;
 import java.security.GeneralSecurityException;
 
 /**
- * This handler authenticates the client credentials : it uses them to get the user profile returned by the provider
- * for an authenticated user.
+ * Specialized handler which builds the authenticated user directly from the retrieved user profile.
  *
  * @author Jerome Leleu
  * @since 3.5.0
  */
-@SuppressWarnings("unchecked")
-public final class ClientAuthenticationHandler extends AbstractPreAndPostProcessingAuthenticationHandler {
-
-    /** Factory to create the principal type. **/
-    @NotNull
-    private PrincipalFactory principalFactory = new DefaultPrincipalFactory();
-
-    /**
-     * The clients for authentication.
-     */
-    @NotNull
-    private final Clients clients;
+public class ClientAuthenticationHandler extends AbstractClientAuthenticationHandler {
 
     /**
      * Whether to use the typed identifier (by default) or just the identifier.
@@ -74,53 +48,29 @@ public final class ClientAuthenticationHandler extends AbstractPreAndPostProcess
      * @param theClients The clients for authentication
      */
     public ClientAuthenticationHandler(final Clients theClients) {
-        this.clients = theClients;
+        super(theClients);
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
-    public boolean supports(final Credential credential) {
-        return credential != null && ClientCredential.class.isAssignableFrom(credential.getClass());
-    }
-
-    @Override
-    protected HandlerResult doAuthentication(final Credential credential) throws GeneralSecurityException, PreventedException {
-        final ClientCredential clientCredentials = (ClientCredential) credential;
-        logger.debug("clientCredentials : {}", clientCredentials);
-
-        final String clientName = clientCredentials.getCredentials().getClientName();
-        logger.debug("clientName : {}", clientName);
-
-        // get client
-        final Client<Credentials, UserProfile> client = this.clients.findClient(clientName);
-        logger.debug("client : {}", client);
-
-        // web context
-        final ServletExternalContext servletExternalContext = (ServletExternalContext) ExternalContextHolder.getExternalContext();
-        final HttpServletRequest request = (HttpServletRequest) servletExternalContext.getNativeRequest();
-        final HttpServletResponse response = (HttpServletResponse) servletExternalContext.getNativeResponse();
-        final WebContext webContext = new J2EContext(request, response);
-        
-        // get user profile
-        final UserProfile userProfile = client.getUserProfile(clientCredentials.getCredentials(), webContext);
-        logger.debug("userProfile : {}", userProfile);
-
-        if (userProfile != null) {
-            final String id;
-            if (typedIdUsed) {
-                id = userProfile.getTypedId();
-            } else {
-                id = userProfile.getId();
-            }
-            if (StringUtils.isNotBlank(id)) {
-              clientCredentials.setUserProfile(userProfile);
-              return new HandlerResult(
-                      this,
-                      new BasicCredentialMetaData(credential),
-                      this.principalFactory.createPrincipal(id, userProfile.getAttributes()));
-            }
+    protected HandlerResult createResult(final ClientCredential credentials, final UserProfile profile)
+        throws GeneralSecurityException, PreventedException {
+        final String id;
+        if (typedIdUsed) {
+            id = profile.getTypedId();
+        } else {
+            id = profile.getId();
         }
-
-        throw new FailedLoginException("Provider did not produce profile for " + clientCredentials);
+        if (StringUtils.isNotBlank(id)) {
+            credentials.setUserProfile(profile);
+            return new HandlerResult(
+                this,
+                new BasicCredentialMetaData(credentials),
+                this.principalFactory.createPrincipal(id, profile.getAttributes()));
+        }
+        throw new FailedLoginException("No identifier found for this user profile: " + profile);
     }
 
     public boolean isTypedIdUsed() {
@@ -129,9 +79,5 @@ public final class ClientAuthenticationHandler extends AbstractPreAndPostProcess
 
     public void setTypedIdUsed(final boolean typedIdUsed) {
         this.typedIdUsed = typedIdUsed;
-    }
-    
-    public void setPrincipalFactory(final PrincipalFactory principalFactory) {
-        this.principalFactory = principalFactory;
     }
 }
